Add docs for OpenID API proxy and general reorg of auth docs (#297)
* Add docs for OpenID API proxy and general reorg of auth docs

This is adding documentation for kiali/kiali#3142. However, since the authentication-related docs are now quite large, this is also re-organizing that content.

All authentication-related content was mostly placed in the installation guide, which is already quite large and somewhat confusing. So, authentication-related content was extracted from the installation guide and placed on it's own pages.

RBAC documentation is also extracted and unified to it's own pages. It is reworked because were scattered and some parts of it were quite outdated.

Related to kiali/kiali#3042
Fixes kiali/kiali#2771

* Feedback: mazz

Co-authored-by: John Mazzitelli <mazz@redhat.com>

Co-authored-by: John Mazzitelli <mazz@redhat.com>